+++ sdk/messaging/eventgrid/aznamespaces/tsp-location.yaml
