+++ sdk/messaging/eventgrid/aznamespaces/tspconfig.yaml
